// src/apispec/src/models/ContextStatsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ResourceUsage } from './ResourceUsage';
import {
    ResourceUsageFromJSON,
    ResourceUsageFromJSONTyped,
    ResourceUsageToJSON,
    ResourceUsageToJSONTyped,
} from './ResourceUsage';

/**
 * 
 * @export
 * @interface ContextStatsResponse
 */
export interface ContextStatsResponse {
    /**
     * 
     * @type {string}
     * @memberof ContextStatsResponse
     */
    contextId?: string;
    /**
     * 
     * @type {string}
     * @memberof ContextStatsResponse
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof ContextStatsResponse
     */
    language?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ContextStatsResponse
     */
    running?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ContextStatsResponse
     */
    paused?: boolean;
    /**
     * 
     * @type {ResourceUsage}
     * @memberof ContextStatsResponse
     */
    usage?: ResourceUsage;
}

/**
 * Check if a given object implements the ContextStatsResponse interface.
 */
export function instanceOfContextStatsResponse(value: object): value is ContextStatsResponse {
    return true;
}

export function ContextStatsResponseFromJSON(json: any): ContextStatsResponse {
    return ContextStatsResponseFromJSONTyped(json, false);
}

export function ContextStatsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContextStatsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'contextId': json['context_id'] == null ? undefined : json['context_id'],
        'type': json['type'] == null ? undefined : json['type'],
        'language': json['language'] == null ? undefined : json['language'],
        'running': json['running'] == null ? undefined : json['running'],
        'paused': json['paused'] == null ? undefined : json['paused'],
        'usage': json['usage'] == null ? undefined : ResourceUsageFromJSON(json['usage']),
    };
}

export function ContextStatsResponseToJSON(json: any): ContextStatsResponse {
    return ContextStatsResponseToJSONTyped(json, false);
}

export function ContextStatsResponseToJSONTyped(value?: ContextStatsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'context_id': value['contextId'],
        'type': value['type'],
        'language': value['language'],
        'running': value['running'],
        'paused': value['paused'],
        'usage': ResourceUsageToJSON(value['usage']),
    };
}