// src/apispec/src/models/ResizeContextRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResizeContextRequest
 */
export interface ResizeContextRequest {
    /**
     * 
     * @type {number}
     * @memberof ResizeContextRequest
     */
    rows: number;
    /**
     * 
     * @type {number}
     * @memberof ResizeContextRequest
     */
    cols: number;
}

/**
 * Check if a given object implements the ResizeContextRequest interface.
 */
export function instanceOfResizeContextRequest(value: object): value is ResizeContextRequest {
    if (!('rows' in value) || value['rows'] === undefined) return false;
    if (!('cols' in value) || value['cols'] === undefined) return false;
    return true;
}

export function ResizeContextRequestFromJSON(json: any): ResizeContextRequest {
    return ResizeContextRequestFromJSONTyped(json, false);
}

export function ResizeContextRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResizeContextRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'rows': json['rows'],
        'cols': json['cols'],
    };
}

export function ResizeContextRequestToJSON(json: any): ResizeContextRequest {
    return ResizeContextRequestToJSONTyped(json, false);
}

export function ResizeContextRequestToJSONTyped(value?: ResizeContextRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'rows': value['rows'],
        'cols': value['cols'],
    };
}